handle all the different kinds of mentions, fix specs
