+++ modules/interaction-handlers.js
@@ -1,11 +1,11 @@
 const responseMessages = require('./response-messages.js');
 const queries = require('../database/queries.js');
 const { MessageAttachment } = require('discord.js');
-const STOP_WORDS = require('../modules/stop-words.js');
 const wordcloudConstructor = require('../modules/wordcloud-constructor.js');
 const { JSDOM } = require('jsdom');
 const canvas = require('canvas');
 const constants = require('./constants.js');
+const { validateAddCommand, formatQuote, mapQuotesToFrequencies } = require('./utilities.js');
 
 module.exports = {
 
@@ -32,7 +32,7 @@ module.exports = {
         try {
             const allQuotesFromServer = await queries.fetchAllQuotes(interaction.guildId);
             if (allQuotesFromServer.length === 0) {
-                await interaction.reply('There haven\'t been any quotes saved from this server, so I didn\'t attach a file.');
+                await interaction.followUp('There haven\'t been any quotes saved from this server, so I didn\'t attach a file.');
                 return;
             }
             for (const quote of allQuotesFromServer) {
@@ -54,7 +54,7 @@ module.exports = {
             });
         } catch (e) {
             console.error(e);
-            await interaction.reply({ content: responseMessages.GENERIC_ERROR, ephemeral: true });
+            await interaction.followUp({ content: responseMessages.GENERIC_ERROR, ephemeral: true });
         }
     },
 
@@ -248,112 +248,3 @@ module.exports = {
         }
     }
 };
-
-async function formatQuote (
-    quote,
-    includeDate = true,
-    includeIdentifier = false,
-    includeMarkdown = true,
-    toFile = false,
-    guildManager = null,
-    interaction = null
-) {
-    const quoteCharacters = ['"', '“', '”'];
-    let quoteMessage = quote.quotation;
-    const d = new Date(quote.said_at);
-
-    if (!quoteCharacters.includes(quoteMessage.charAt(0))) {
-        quoteMessage = '"' + quoteMessage;
-    }
-
-    if (!quoteCharacters.includes(quoteMessage.charAt(quoteMessage.length - 1))) {
-        quoteMessage = quoteMessage + '"';
-    }
-
-    if (includeMarkdown) {
-        quoteMessage = '_' + quoteMessage + '_';
-    }
-
-    if (toFile && /^<@[&#!0-9]+>$/.test(quote.author)) { // Discord @s are represented as <@UserID>
-        quoteMessage = quoteMessage + ' - ' + await attemptToResolveMentionToNickname(guildManager, interaction, quote.author);
-    } else {
-        quoteMessage = quoteMessage + ' - ' + quote.author;
-    }
-
-    if (includeDate) {
-        quoteMessage += ' (added ' + d.toLocaleString('default', {
-            year: 'numeric',
-            month: 'short',
-            day: 'numeric'
-        }) + ')';
-    }
-
-    if (includeIdentifier) {
-        quoteMessage += ' (**identifier**: ' + quote.id + ')';
-    }
-
-    return quoteMessage;
-}
-
-/* When quotes are written to a text file, we want to resolve a Users ID to their nickname in the server, if it still exists,
-    so that we don't end up writing a representation of the mention (e.g. <@123>) as the author. */
-async function attemptToResolveMentionToNickname (guildManager, interaction, author) {
-    try {
-        const guild = await guildManager.fetch(interaction.guildId);
-        if (guild && guild.members) {
-            const member = await guild.members.fetch(author.replace(/[^0-9]/g, ''));
-            if (member) {
-                return member.nickname
-                    ? '@' + member.nickname
-                    : '@' + member.user.username;
-            }
-        }
-        return 'User not found';
-    } catch (e) {
-        return 'User not found';
-    }
-}
-
-async function validateAddCommand (quote, author, interaction) {
-    let reply = 'Your quote has the following problems:\n\n';
-    let hasProblem = false;
-    if (quote.length > constants.MAX_QUOTE_LENGTH) {
-        reply += '- Your quote of length ' + quote.length + ' characters exceeds the maximum allowed length of ' +
-                constants.MAX_QUOTE_LENGTH + ' characters.\n';
-        hasProblem = true;
-    }
-    if (author.length > constants.MAX_AUTHOR_LENGTH) {
-        reply += '- Your author of length ' + author.length + ' characters exceeds the maximum allowed length of ' +
-                constants.MAX_AUTHOR_LENGTH + ' characters.\n';
-        hasProblem = true;
-    }
-    if (quote.toLowerCase().includes('http://') || quote.toLowerCase().includes('https://')) {
-        reply += '- Quotes with links are disallowed.';
-        hasProblem = true;
-    }
-    if (hasProblem) {
-        await interaction.reply({ content: reply, ephemeral: true });
-    }
-}
-
-function mapQuotesToFrequencies (quotesForCloud) {
-    const wordsWithOccurrences = [];
-    for (const quote of quotesForCloud) {
-        const words = quote.quotation
-            .split(' ')
-            .map((word) => word.toLowerCase().replace(/[^a-zA-Z0-9-']/g, ''))
-            .filter((word) => word.length > 0 && !STOP_WORDS.includes(word));
-        for (const word of words) {
-            const existingWord = wordsWithOccurrences.find((element) => element.word === word);
-            if (existingWord) {
-                existingWord.frequency ++;
-            } else {
-                wordsWithOccurrences.push({
-                    word,
-                    frequency: 1
-                });
-            }
-        }
-    }
-    return wordsWithOccurrences;
-}
